+++ examples/update_bids_datasets.py
@@ -5,9 +5,9 @@
 
 When working with electrophysiological data in the BIDS format, we usually
 do not have all the metadata stored in the ``Raw`` mne-python object.
-We can update the BIDS sidecar files via the ``update_sidecars`` function.
+We can update the BIDS sidecar files via the ``update_sidecar_json`` function.
 
-In this tutorial, we show how ``update_sidecars`` can be used to update and
+In this tutorial, we show how ``update_sidecar_json`` can be used to update and
 modify BIDS-formatted data.
 
 """
@@ -28,15 +28,14 @@ from mne.datasets import somato
 import tempfile
 import json
 from mne_bids import (BIDSPath, read_raw_bids,
-                      print_dir_tree, make_report, update_sidecars)
+                      print_dir_tree, make_report, update_sidecar_json)
 
 ###############################################################################
 # We will be using the `MNE somato data <mne_somato_data_>`_, which
 # is already stored in BIDS format.
 # For more information, you can check out the
 # respective :ref:`example <ex-convert-mne-sample>`.
-
-###############################################################################
+#
 # Download the ``somato`` BIDS dataset
 # ------------------------------------
 #
@@ -60,8 +59,8 @@ print_dir_tree(bids_root, max_depth=3)
 print(make_report(bids_root))
 
 ###############################################################################
-# Update the dataset contents
-# ---------------------------
+# Update the sidecar JSON dataset contents
+# ----------------------------------------
 #
 # We can use MNE-BIDS to update all sidecar files for a matching
 # ``BIDSPath`` object. We then pass in a sidecar template to update
@@ -76,7 +75,8 @@ task = 'somato'
 suffix = 'meg'
 
 bids_path = BIDSPath(subject=subject, task=task, suffix=suffix,
-                     datatype=datatype, root=bids_root)
+                     datatype=datatype, extension='.json',
+                     root=bids_root)
 
 # We can now retrieve a list of all MEG-related files in the dataset:
 print(bids_path.match())
@@ -89,7 +89,7 @@ sidecar_template = {
 }
 
 # Now update all sidecar fields according to the template
-update_sidecars(bids_path=bids_path, sidecar_template=sidecar_template)
+update_sidecar_json(bids_path=bids_path, sidecar_template=sidecar_template)
 
 ###############################################################################
 # Read the updated dataset
@@ -124,12 +124,12 @@ with tempfile.TemporaryDirectory() as tempdir:
         json.dump(sidecar_template, fout)
 
     # Update sidecar files via a template defined as a JSON file.
-    update_sidecars(bids_path=bids_path,
-                    sidecar_template=sidecar_template_fpath)
+    update_sidecar_json(bids_path=bids_path,
+                        sidecar_template=sidecar_template_fpath)
 
 ###############################################################################
 # Now let us inspect the dataset again by generating the report again. Now that
-# ``update_sidecars`` was called, the metadata will be updated.
+# ``update_sidecar_json`` was called, the metadata will be updated.
 
 # The power line frequency should now change back
 raw = read_raw_bids(bids_path=bids_path)