+++ mne_bids/sidecar_updates.py
@@ -6,24 +6,58 @@
 import json
 from collections import OrderedDict
 
-from mne.utils import warn
-
 from mne_bids.utils import _write_json
+
+# XXX: should we encode this, or let bids-validator do its job?
+# def _check_specification_compat(datatype):
+#     SIDECAR_JSON_FIELDS = {
+#         'meg': ['TaskName', 'InstitutionName', 'InstitutionAddress', 'Manufacturer',
+#                 'ManufacturersModelName', 'SoftwareVersions', 'TaskDescription',
+#                 'Instructions', 'CogAtlasID', 'CogPOID', 'DeviceSerialNumber',
+#                 'SamplingFrequency', 'PowerLineFrequency',
+#                 'SoftwareFilters', 'HardwareFilters',
+#                 'DigitizedLandmarks', 'DigitizedHeadPoints',
+#                 'MEGChannelCount', 'MEGREFChannelCount',
+#                 'EEGChannelCount', 'ECOGChannelCount',
+#                 'SEEGChanelCount', 'EOGChannelCount',
+#                 'EMGChannelCount', 'MiscChannelCount',
+#                 'TriggerChannelCount', 'RecordingDuration',
+#                 'RecordingType', 'EpochLength',
+#                 'ContinuousHeadLocalization', 'HeadCoilFrequency',
+#                 'MaxMovement', 'SubjectArtefactDescription',
+#                 'AssociatedEmptyRoom', 'EEGPlacementScheme',
+#                 'CapManufacturer', 'CapManufacturersModelName', 'EEGReference'
+#                 'DewarPosition', ],
+#         'coordsystem': [f'{datatype}CoordinateSystem', f'{datatype}CoordinateUnits',
+#                         f'{datatype}CoordinateSystemDescription',
+#                         'HeadCoilCoordinates', 'HeadCoilCoordinateSystem',
+#                         'HeadCoilCoordinateUnits', 'HeadCoilCoordinateSystemDescription',
+#                         'DigitizedHeadPoints', 'DigitizedHeadPointsCoordinateSystem',
+#                         'DigitizedHeadPointsCoordinateUnits',
+#                         'DigitizedHeadPointsCoordinateSystemDescription',
+#                         ]
+#     }
+
 
 
 # TODO: add support for tsv files
-def update_sidecars(bids_path, sidecar_template, verbose=True):
+def update_sidecar_json(bids_path, sidecar_template, verbose=True):
     """Update sidecar files using template JSON file.
 
-    Will match all paths based on BIDS entities defined in the ``bids_path``
-    argument and update metadata fields according to the ``sidecar_template``.
-    If a field does not exist in the corresponding sidecar file, then that
-    field will not be updated. For example, if ``InstitutionName`` is
-    not defined in the sidecar json file, then trying to update
-    ``InstitutionName`` to ``Martinos Center`` will not update
-    the sidecar json file.
+    Will update metadata fields inside the path defined by
+    ``bids_path.fpath`` according to the ``sidecar_template``.
+    If a field does not exist in the corresponding sidecar file,
+    then that field will be created according to the template.
+    If a field does exist in the corresponding sidecar file,
+    then that field will be updated according to the template.
 
-    Currently, only works for JSON files.
+    For example, if ``InstitutionName`` is
+    not defined in the sidecar json file, then trying to update
+    ``InstitutionName`` to ``Martinos Center`` will update
+    the sidecar json file to have ``InstitutionName`` as
+    ``Martinos Center``.
+
+    Note that
 
     Parameters
     ----------
@@ -33,90 +67,82 @@ def update_sidecars(bids_path, sidecar_template, verbose=True):
         ``.datatype`` attribute **may** be set. If ``.datatype`` is
         not set and only one data type (e.g., only EEG or MEG data)
         is present in the dataset, it will be
-        selected automatically. All matching bids paths via
-        ``bids_path.match()`` will be updated according to
-        the ``sidecar_template``.
+        selected automatically. This must uniquely identify
+        an existing file path, else an error will be raised.
     sidecar_template : dict | str | pathlib.Path
         A dictionary, or JSON file that defines the
         sidecar fields and corresponding values to be updated to.
     verbose : bool
         The verbosity level.
 
+    Notes
+    -----
+    This function only works for JSON files.
+
+    Sidecar JSON files include files such as ``*_ieeg.json``,
+    ``*_coordsystem.json``, ``*_scans.json``, etc.
+
+    You should double check that your template is correct for the
+    corresponding sidecar JSON file because it will perform
+    a dictionary update of the sidecar fields according to
+    the passed in template overwriting any information that was
+    previously there.
+
+    Raises
+    ------
+    RuntimeError
+        If the specified ``bids_path.fpath`` cannot be found
+        in the dataset.
+
     Examples
     --------
-    >>> # update all files in BIDS dataset
-    >>> bids_path = BIDSPath(root='./')
-    >>> update_sidecars(bids_path, sidecar_template_fpath)
+    >>> # update sidecar json file
+    >>> bids_path = BIDSPath(root='./', subject='001', session='001',
+                             task='test', run='01', suffix='ieeg',
+                             extension='.json')
+    >>> sidecar_template = {'PowerLineFrequency': 50}
+    >>> update_sidecar_json(bids_path, sidecar_template)
+    >>> # update sidecar coordsystem json file
+    >>> bids_path = BIDSPath(root='./', subject='001', session='001',
+                             suffix='coordsystem', extension='.json')
+    >>> sidecar_template = {'iEEGCoordinateSyste,': 'Other'}
+    >>> update_sidecar_json(bids_path, sidecar_template)
     """
     # get all matching json files
     bids_path = bids_path.copy()
-    bids_path.update(extension='.json')
-    sidecar_paths = bids_path.match(return_json=True)
+    if bids_path.extension != '.json':
+        raise RuntimeError('Only works for ".json" files. The '
+                           'BIDSPath object passed in has '
+                           f'{bids_path.extension} extension.')
 
+    # get the file path
+    fpath = bids_path.fpath
+    if not fpath.exists():
+        raise RuntimeError(f'Sidecar file {fpath} does not '
+                           f'exist.')
+
+    # sidecar template either from file, or as dictionary
     if isinstance(sidecar_template, dict):
         sidecar_tmp = sidecar_template
     else:
-        # load sidecar template
         with open(sidecar_template, 'r') as tmp_f:
-            sidecar_tmp = json.load(tmp_f,
-                                    object_pairs_hook=OrderedDict
-                                    )
-
-    template_fields = set(sidecar_tmp.keys())
+            sidecar_tmp = json.load(
+                tmp_f, object_pairs_hook=OrderedDict)
 
     if verbose:
         print(sidecar_tmp)
-        print(f'Updating {sidecar_paths}...')
+        print(f'Updating {fpath}...')
 
-    # keep track of all used fields in the template
-    used_fields = []
+    # load in sidecar filepath
+    with open(fpath, 'r') as tmp_f:
+        sidecar_json = json.load(
+            tmp_f, object_pairs_hook=OrderedDict)
 
-    # update all matching sidecar paths
-    for sidecar_bids_path in sidecar_paths:
-        fpath = sidecar_bids_path.fpath
+    # update sidecar JSON file with the template fields
+    sidecar_json.update(**sidecar_tmp)
 
-        # load in sidecar filepath
-        with open(fpath, 'r') as tmp_f:
-            sidecar_json = json.load(
-                tmp_f, object_pairs_hook=OrderedDict)
-
-        # get the fields inside JSON
-        default_fields = set(sidecar_json.keys())
-
-        # XXX: how to we do this if fields are not in the sidecar file already?
-
-        # get fields that are not in sidecar file
-        # only_in_template = template_fields.difference(default_fields)
-        # get fields that are in both the sidecar file and template
-        in_both = template_fields.intersection(default_fields)
-
-        # keep track of unused fields
-        for key in in_both:
-            used_fields[key] = 1
-
-        # Use field order in template to sort default keys, if values are None
-        field_order = list(sidecar_tmp.keys())
-
-        # pop fields in the template if they are overwriting value to None
-        for field in in_both:
-            if sidecar_tmp[field] is None:
-                sidecar_tmp.pop(field)
-
-        # Update values in generate sidecar with non-None values in template
-        sidecar_json.update(sidecar_tmp)
-
-        # Sort updated sidecar according to sort order in template
-        sorted_info = [(field, sidecar_json[field]) for field in field_order]
-        sidecar_json = OrderedDict(sorted_info)
-
-        _write_json(fpath, sidecar_json, overwrite=True, verbose=verbose)
-
-    # warn user that there are some unused fields in the sidecar template
-    # that need to first be instantiated in the file before it can update
-    unused_fields = [key for key in sidecar_tmp.keys()
-                     if key not in used_fields]
-    warn('Leftover template fields were not used to update any '
-         f'sidecar file {unused_fields}')
+    # write back the sidecar JSON
+    _write_json(fpath, sidecar_json, overwrite=True, verbose=verbose)
 
 
 def _update_sidecar(sidecar_fname, key, val):